Add files via upload
This is my first FiveM script, feel free to do whatever you want with it except sell it. I've only tested it with one player but i will test it by adding fake players to it and see so it works, if it doesnt work i will fix it and update it.